langchain-demo/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-postgres
python-dotenv

